General updates

+++ backend/flaskr/__init__.py
@@ -4,6 +4,7 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import func
 from flask_cors import CORS
 import random
+import collections
 
 from models import setup_db, Question, Category
 
@@ -60,25 +61,28 @@ def create_app(test_config=None):
     '''
     @app.route('/api/v1/questions')
     def get_questions():
-        try:
-            page = request.args.get('page', default=1, type=int)
-            per_page = request.args.get('per_page', default=10, type=int)
-            questions_query = Question.query.paginate(page=page, error_out=False, max_per_page=per_page)
-            formatted_questions = [question.format() for question in questions_query.items]
+        page = request.args.get('page', default=1, type=int)
+        per_page = request.args.get('per_page', default=10, type=int)
 
-            categories = Category.query.all()
-            formatted_categories = {}
-            for category in categories:
-                formatted_categories[category.id] = category.type
+        # page must be larger than zero and per_page must be less or equal to one hundred
+        if page < 1 or per_page > 100:
+            return abort(422)
 
-            return jsonify({
-                'questions': formatted_questions,
-                'total_questions': questions_query.total,
-                'categories': formatted_categories,
-                'current_category': None
-            })
-        except:
-            abort(404)
+        questions_query = Question.query.paginate(page=page, error_out=False, max_per_page=per_page)
+        formatted_questions = [question.format() for question in questions_query.items]
+
+        categories = Category.query.all()
+        formatted_categories = {}
+        for category in categories:
+            formatted_categories[category.id] = category.type
+
+        return jsonify({
+            'success': True,
+            'questions': formatted_questions,
+            'total_questions': questions_query.total,
+            'categories': formatted_categories,
+            'current_category': None
+        })
 
     '''
     Create an endpoint to DELETE question using a question ID.
@@ -88,18 +92,16 @@ def create_app(test_config=None):
     '''
     @app.route('/api/v1/questions/<int:question_id>', methods=['DELETE'])
     def delete_question(question_id):
-        try:
-            question = Question.query.get(question_id)
-            question.delete()
+        question = Question.query.filter(Question.id == question_id).one_or_none()
 
-            return jsonify({
-                'success': True
-            })
-        except:
-            db.session.rollback()
-            abort(404)
-        finally:
-            db.session.close()
+        # question must exist
+        if question is None:
+            return abort(404)
+
+        question.delete()
+        return jsonify({
+            'success': True
+        })
 
     '''
     Create an endpoint to POST a new question, 
@@ -112,24 +114,25 @@ def create_app(test_config=None):
     '''
     @app.route('/api/v1/questions', methods=['POST'])
     def create_question():
-        try:
-            request_body = request.get_json()
-            question = Question(
-                request_body['question'],
-                request_body['answer'],
-                request_body['category'],
-                request_body['difficulty']
-            )
-            question.insert()
+        required_request_data = ['question', 'answer', 'difficulty', 'category']
+        request_body = request.get_json()
 
-            return jsonify({
-                'success': True
-            })
-        except:
-            db.session.rollback()
-            abort(500)
-        finally:
-            db.session.close()
+        for key in required_request_data:
+            # all required request data must exist
+            if key not in request_body or request_body[key] is None or request_body[key] == '':
+                return abort(422)
+
+        question = Question(
+            request_body['question'],
+            request_body['answer'],
+            request_body['category'],
+            request_body['difficulty']
+        )
+        question.insert()
+
+        return jsonify({
+            'success': True
+        })
 
     '''
     POST endpoint to get questions based on a search term. 
@@ -138,19 +141,17 @@ def create_app(test_config=None):
     '''
     @app.route('/api/v1/questions/searches', methods=['POST'])
     def search_questions():
-        try:
-            request_body = request.get_json()
-            search_term = request_body['search_term']
-            questions = Question.query.filter(Question.question.ilike(f'%{search_term}%')).all()
-            formatted_questions = [question.format() for question in questions]
+        request_body = request.get_json()
+        search_term = request_body['search_term']
+        questions = Question.query.filter(Question.question.ilike(f'%{search_term}%')).all()
+        formatted_questions = [question.format() for question in questions]
 
-            return jsonify({
-                'questions': formatted_questions,
-                'total_questions': len(formatted_questions),
-                'current_category': None
-            })
-        except:
-            abort(404)
+        return jsonify({
+            'success': True,
+            'questions': formatted_questions,
+            'total_questions': len(formatted_questions),
+            'current_category': None
+        })
 
     '''
     Create a GET endpoint to get questions based on category. 
@@ -161,18 +162,21 @@ def create_app(test_config=None):
     '''
     @app.route('/api/v1/categories/<int:category_id>/questions')
     def get_category_questions(category_id):
-        try:
-            category = Category.query.get(category_id)
-            questions = Question.query.filter_by(category=category.id).all()
-            formatted_questions = [question.format() for question in questions]
+        category = Category.query.filter(Category.id == category_id).one_or_none()
 
-            return jsonify({
-                'questions': formatted_questions,
-                'total_questions': len(questions),
-                'current_category': category.format()
-            })
-        except:
-            abort(404)
+        # category must exist
+        if category is None:
+            return abort(422)
+
+        questions = Question.query.filter_by(category=str(category.id)).all()
+        formatted_questions = [question.format() for question in questions]
+
+        return jsonify({
+            'success': True,
+            'questions': formatted_questions,
+            'total_questions': len(questions),
+            'current_category': category.format()
+        })
 
     '''
     Create a POST endpoint to get questions to play the quiz. 
@@ -186,51 +190,48 @@ def create_app(test_config=None):
     '''
     @app.route('/api/v1/quizzes', methods=['POST'])
     def play_quiz():
-        try:
-            request_body = request.get_json()
-            previous_questions = request_body['previous_questions']
+        request_body = request.get_json()
 
-            if 'quiz_category' in request_body and not (request_body['quiz_category'] is None):
-                quiz_category = request_body['quiz_category']
-                questions = Question.query.filter_by(category=int(quiz_category['id']))
-            else:
-                questions = Question.query.all()
+        # request body must exist
+        if request_body is None:
+            return abort(422)
 
-            not_taken_questions = []
+        # request body must contain 'previous_questions' field with iterable value
+        if not ('previous_questions' in request_body) or not isinstance(request_body['previous_questions'], collections.Iterable):
+            return abort(422)
 
-            for question in questions:
-                already_taken = False
+        previous_questions = request_body['previous_questions']
 
-                for previous_question_id in previous_questions:
-                    if previous_question_id == question.id:
-                        already_taken = True
-                        break
+        if 'quiz_category' in request_body and not (request_body['quiz_category'] is None):
+            quiz_category = request_body['quiz_category']
+            not_taken_questions = Question.query.filter(
+                Question.category == str(quiz_category['id']),
+                Question.id.notin_(previous_questions)
+            ).all()
+        else:
+            not_taken_questions = Question.query.filter(Question.id.notin_(previous_questions)).all()
 
-                if not already_taken:
-                    not_taken_questions.append(question)
+        random_not_taken_question = random.choice(not_taken_questions)
 
-            random_not_taken_question = random.choice(not_taken_questions)
-
-            return jsonify({
-                'question': random_not_taken_question.format()
-            })
-        except:
-            abort(404)
+        return jsonify({
+            'success': True,
+            'question': random_not_taken_question.format()
+        })
 
     '''
     Create error handlers for all expected errors 
     including 404 and 422. 
     '''
     @app.errorhandler(400)
-    def bad_request():
+    def bad_request(error):
         return jsonify({
             'success': False,
             'error': 400,
             'message': 'Bad request'
-        })
+        }), 400
 
     @app.errorhandler(404)
-    def not_found():
+    def not_found(error):
         return jsonify({
             'success': False,
             'error': 404,
@@ -238,27 +239,27 @@ def create_app(test_config=None):
         }), 404
 
     @app.errorhandler(422)
-    def unprocessable_entity():
+    def unprocessable_entity(error):
         return jsonify({
             'success': False,
             'error': 422,
             'message': 'Unprocessable entity'
-        })
+        }), 422
 
     @app.errorhandler(405)
-    def method_not_allowed():
+    def method_not_allowed(error):
         return jsonify({
             'success': False,
             'error': 405,
             'message': 'Method not allowed'
-        })
+        }), 405
 
     @app.errorhandler(500)
-    def internal_server_error():
+    def internal_server_error(error):
         return jsonify({
             'success': False,
             'error': 500,
             'message': 'Internal server error'
-        })
+        }), 500
 
     return app
